requirements.txt
Flask
flask-cors
google-generativeai
langchain-google-genai
langchain-community
langchain-core
langchain
chromadb
